Time: 50 ms (52.19%) | Memory: 28.9 MB (99.03%) - LeetSync

+++ 15-3sum/3sum.cpp
@@ -7,17 +7,17 @@ public:
         sort(nums.begin(),nums.end());
 
         for(int i=0; i<n-2; i++){
-            if (i > 0 && nums[i] == nums[i - 1]) continue; // skip duplicate
+            if (i > 0 && nums[i] == nums[i - 1]) continue; 
             int left = i+1;
             int right = n-1;
 
             while(left<right){
-                long long sum =(long long) nums[i] + nums[left] + nums[right];
+                long long sum = nums[i] + nums[left] + nums[right];
                 if(sum == 0){
                     res.push_back({nums[i],nums[left],nums[right]});
                     left++;
                     right--;
-                     // skip duplicates on both sides
+
                     while (left < right && nums[left] == nums[left - 1]) left++;
                     while (left < right && nums[right] == nums[right + 1]) right--;
                 }
